Add choosing assistance systems

+++ src/main/java/be/uhasselt/dwi_application/model/Jackson/DisplayConfig.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 public class DisplayConfig {
-    private final int id;
+    private final Long id;
     private final int quantity;
     private final int brightness;
     private final DisplayStatus status;
@@ -18,7 +18,7 @@ public class DisplayConfig {
         }
     }
 
-    public DisplayConfig(int id, int number, int brightness, DisplayStatus status) {
+    public DisplayConfig(Long id, int number, int brightness, DisplayStatus status) {
         this.id = id;
         this.quantity = number;
         this.status = status;
@@ -29,15 +29,15 @@ public class DisplayConfig {
         this.brightness = brightness;
     }
 
-    public static DisplayConfig off(int id) {
+    public static DisplayConfig off(Long id) {
         return new DisplayConfig(id, 0, 0, DisplayStatus.OFF);
     }
 
-    public static DisplayConfig on(int id, int quantity) {
+    public static DisplayConfig on(Long id, int quantity) {
         return new DisplayConfig(id, quantity, 7, DisplayStatus.ON);
     }
 
-    public int getId() {return id;}
+    public Long getId() {return id;}
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     public Integer getQuantity() {return status == DisplayStatus.OFF ? null : quantity;}
